mid point in moving every modifiers
