+++ lib/period.dart
@@ -21,7 +21,7 @@ import 'src/period.dart';
 import 'src/period_generator.dart';
 
 export 'src/enums.dart' show Month, PeriodGenerator, Week, Weekday;
-export 'src/every_modifier.dart';
+export 'src/everies/everies.dart';
 export 'src/extensions.dart' show EndOfDay;
 export 'src/period.dart';
 export 'src/period_generator.dart';
